# server/docs/swagger.yaml
definitions:
...
  eCommerce.Shop:
    properties:
      ID:
        description: 主键ID
        type: integer
      clientCert:
        description: API Client Cert
        type: string
      clientId:
        description: API Client ID
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      deletedBy:
        type: integer
      platformName:
        description: 平台名称
        type: string
      shopId:
        description: 平台ID
        type: string
      shopName:
        description: 店铺名称
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.6.1
paths:
...
  /shop/createShop:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建店铺
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/eCommerce.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建店铺
      tags:
      - Shop
  /shop/deleteShop:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除店铺
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/eCommerce.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除店铺
      tags:
      - Shop
  /shop/deleteShopByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除店铺
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除店铺
      tags:
      - Shop
  /shop/findShop:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: API Client Cert
        in: query
        name: clientCert
        type: string
      - description: API Client ID
        in: query
        name: clientId
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - in: query
        name: deletedBy
        type: integer
      - description: 平台名称
        in: query
        name: platformName
        type: string
      - description: 平台ID
        in: query
        name: shopId
        type: string
      - description: 店铺名称
        in: query
        name: shopName
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询店铺
      tags:
      - Shop
  /shop/getShopList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: API Client Cert
        in: query
        name: clientCert
        type: string
      - description: API Client ID
        in: query
        name: clientId
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - in: query
        name: deletedBy
        type: integer
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 平台名称
        in: query
        name: platformName
        type: string
      - description: 平台ID
        in: query
        name: shopId
        type: string
      - description: 店铺名称
        in: query
        name: shopName
        type: string
      - in: query
        name: startCreatedAt
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取店铺列表
      tags:
      - Shop
  /shop/updateShop:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新店铺
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/eCommerce.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新店铺
      tags:
      - Shop
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"